jf_signature

Module bls_over_bls12381

BLS Signature Scheme

Conforming to IRTF draft, wrapping blst crate under the hood.

Examples

use rand_chacha::{ChaCha20Rng, rand_core::SeedableRng};
use jf_signature::{SignatureScheme, bls_over_bls12381::BLSSignatureScheme};

let pp = BLSSignatureScheme::param_gen::<ChaCha20Rng>(None)?;

// make sure the PRNG passed has good and trusted entropy.
// you could use `OsRng` from `rand_core` or `getrandom` crate,
// or a `SeedableRng` like `ChaChaRng` with seed generated from good randomness source.
let mut rng = ChaCha20Rng::from_seed([0u8; 32]);
let (sk, pk) = BLSSignatureScheme::key_gen(&pp, &mut rng)?;

let msg = "The quick brown fox jumps over the lazy dog";
let sig = BLSSignatureScheme::sign(&pp, &sk, &msg, &mut rng)?;
assert!(BLSSignatureScheme::verify(&pp, &pk, &msg, &sig).is_ok());

Generating independent keys from the same IKM

In case you want to keep the IKM for multiple key pairs, and potentially reconstruct them later on from IKM.

use rand_chacha::{ChaCha20Rng, rand_core::{SeedableRng, RngCore}};
use sha2::{Sha256, Digest};
use jf_signature::{SignatureScheme, bls_over_bls12381::BLSSignatureScheme};

let mut rng = ChaCha20Rng::from_seed([0u8; 32]); // seed from proper entropy source in practice!
let pp = BLSSignatureScheme::param_gen::<ChaCha20Rng>(None)?;

// NOTE: in practice, please use [`zeroize`][zeroize] to wipe sensitive
// key materials out of memory.
let mut ikm = [0u8; 32]; // should be at least 32 bytes
rng.fill_bytes(&mut ikm);

let mut hasher = Sha256::new();
hasher.update(b"MY-BLS-SIG-KEYGEN-SALT-DOM-SEP");
let salt = hasher.finalize();

let (sk1, pk1) = BLSSignatureScheme::key_gen_v5(&ikm, &salt, b"banking".as_ref())?;
let (sk2, pk2) = BLSSignatureScheme::key_gen_v5(&ikm, &salt, b"legal".as_ref())?;

let msg = "I authorize transferring 10 dollars to Alice";
let sig = BLSSignatureScheme::sign(&pp, &sk1, &msg, &mut rng)?;
assert!(BLSSignatureScheme::verify(&pp, &pk1, &msg, &sig).is_ok());

let msg = "I agree to the Terms and Conditions.";
let sig = BLSSignatureScheme::sign(&pp, &sk2, &msg, &mut rng)?;
assert!(BLSSignatureScheme::verify(&pp, &pk2, &msg, &sig).is_ok());

Structs

BLSSignKey

A BLS Secret Key (Signing Key). We intentionally omit the implementation of serde::Serializable so that it won’t be unnoticably serialized or printed out through outer structs. However, users could manually serialize it into bytes by calling to_bytes() or to_tagged_base64() and exercise with self-cautions.

BLSSignature

A BLS Signature.

BLSSignatureScheme

BLS signature scheme. Wrapping around structs from the blst crate. See module-level documentation for example usage.

BLSVerKey

A BLS Public Key (Verification Key).