jf_signature

Module bls_over_bn254

This module implements the BLS signature over the BN254 curve. The BLS signature scheme relies on a bilinear map e:G1 x G2 -> GT and a hash function H from a message input space (e.g. bytes) to a group element G1. H must be such that for all m, the discrete logarithm H(m) w.r.t to the generator of G1 is unknown.

The scheme works as follows: Let g1 and g2 be generators of G1 and G2.

KeyGen()

sample a random s in the scalar field Fr and return the key pair (sk,pk):=(s,g2^s)

Sign(sk,m)

return sigma=H(m)^{sk}

Verify(pk,m,sigma)

Check that e(sigma,g_2)=e(H(m),pk)

In this module:

e is the pairing over the curve BN254 supported by the EVM EIP-196, EIP197

H is implemented using the “hash-and-pray” approach. See function hash_to_curve

Structs

BLSOverBN254CurveSignatureScheme

BLS signature scheme.

KeyPair

Signature secret key pair used to sign messages

SignKey

Signing key for BLS signature. We intentionally omit the implementation of serde::Serializable so that it won’t be unnoticably serialized or printed out through outer structs. However, users could manually serialize it into bytes by calling to_bytes() or to_tagged_base64() and exercise with self-cautions.

Signature

The signature of BLS signature scheme

VerKey

Signature public verification key

Functions

hash_to_curve

Non constant time hash to curve algorithm (a.k.a “hash-and-pray”) The hashing algorithm consists of the following steps: