jf_signature

Module constants

Tags to be used for tagged-base64

Modules

tag

Tags

Constants

BLS_SIG_COMPRESSED_PK_SIZE

Size in bytes of a compressed verification key in our BLS signature scheme.

BLS_SIG_COMPRESSED_SIGNATURE_SIZE

Size in bytes of a compressed signature in our BLS signature scheme.

BLS_SIG_PK_SIZE

Size in bytes of a verification key in our BLS signature scheme.

BLS_SIG_SIGNATURE_SIZE

Size in bytes of a signature in our BLS signature scheme.

BLS_SIG_SK_SIZE

Size in bytes of a secret key in our BLS signature scheme.

CS_ID_BLS_BN254

ciphersuite identifier for BLS signature over BN254 Note this is adapted from https://www.ietf.org/archive/id/draft-irtf-cfrg-bls-signature-05.html#name-ciphersuite-format. In particular the “hash-and-pray” method is not part of https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16, so the tag “NCTH” (non constant time hash) is not standard.

CS_ID_BLS_MIN_SIG

ciphersuite identifier for BLS signature over BLS12_381, see: https://www.ietf.org/archive/id/draft-irtf-cfrg-bls-signature-05.html#name-ciphersuite-format

CS_ID_SCHNORR

ciphersuite identifier for schnorr signature