jf_signature

Trait AggregateableSignatureSchemes

pub trait AggregateableSignatureSchemes:
    SignatureScheme
    + Serialize
    + for<'a> Deserialize<'a>
    + Debug {
    // Required methods
    fn aggregate(
        pp: &Self::PublicParameter,
        vks: &[Self::VerificationKey],
        sigs: &[Self::Signature],
    ) -> Result<Self::Signature, SignatureError>;
    fn aggregate_verify<M: AsRef<[Self::MessageUnit]>>(
        pp: &Self::PublicParameter,
        vks: &[Self::VerificationKey],
        msgs: &[M],
        sig: &Self::Signature,
    ) -> Result<(), SignatureError>;
    fn multi_sig_verify(
        pp: &Self::PublicParameter,
        vks: &[Self::VerificationKey],
        msg: &[Self::MessageUnit],
        sig: &Self::Signature,
    ) -> Result<(), SignatureError>;
}

Trait for aggregatable signatures. TODO: generic over hash functions

Required Methods

fn aggregate( pp: &Self::PublicParameter, vks: &[Self::VerificationKey], sigs: &[Self::Signature], ) -> Result<Self::Signature, SignatureError>

Aggregate multiple signatures into a single signature The list of public keys is also in the input as some aggregate signature schemes might also use pks for aggregation

fn aggregate_verify<M: AsRef<[Self::MessageUnit]>>( pp: &Self::PublicParameter, vks: &[Self::VerificationKey], msgs: &[M], sig: &Self::Signature, ) -> Result<(), SignatureError>

Verify an aggregate signature w.r.t. a list of messages and public keys. It is user’s responsibility to ensure that the public keys are validated.

fn multi_sig_verify( pp: &Self::PublicParameter, vks: &[Self::VerificationKey], msg: &[Self::MessageUnit], sig: &Self::Signature, ) -> Result<(), SignatureError>

Verify a multisignature w.r.t. a single message and a list of public keys. It is user’s responsibility to ensure that the public keys are validated.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.